correcting errors on visual studio compiler
