+++ modules/ustk_io/include/visp3/ustk_io/usRfReader.h
@@ -85,82 +85,5 @@ private:
   std::ifstream m_dataFile;
 };
 
-/****************************************************************************
-* Template implementations.
-****************************************************************************/
-
-/**
-* Default constructor.
-*/
-usRfReader::usRfReader() : m_fileName(""), m_fileNameIsSet(false), m_header()
-{
-
-}
-
-/**
-* Destructor.
-*/
-usRfReader::~usRfReader()
-{
-
-}
-
-/**
-* FileName setter.
-* @param fileName Name of the .vol file.
-*/
-void usRfReader::setFileName(const std::string &fileName)
-{
-  m_fileName = fileName;
-  m_fileNameIsSet = true;
-}
-
-/**
-* Sequence opening.
-* @param image First image of the sequence to read.
-*/
-void usRfReader::open(usImageRF2D<short int> &image)
-{
-  if(!m_fileNameIsSet) {
-    throw(vpException(vpException::badValue, "Sequence settings file name not set"));
-  }
-
-  if(usImageIo::getHeaderFormat(m_fileName) != usImageIo::FORMAT_RF)
-    throw(vpException(vpException::ioError, "only .rf files are supported"));
-
-  //INIT
-  int szHeader = sizeof(usImageIo::FrameHeader);
-  int n = 0;
-  char byte;
-
-  //FILE OPENING
-  m_dataFile.open( m_fileName.c_str(), std::ios::in|std::ios::binary);
-
-  //READING HEADER
-  while (n < szHeader) {
-    m_dataFile.read(&byte, 1);
-    ((char*)&m_header)[n] = byte;
-    n++;
-  }
-
-  //CHECK IMAGE TYPE
-  if (m_header.type != 16)
-    throw(vpException(vpException::badValue, "trying to read non-rf data in .rf file"));
-
-  //CHECK DATA TYPE
-  if (m_header.ss != 16)
-    throw(vpException(vpException::badValue, ".vol file doesn't contain short data"));
-
-  image.resize( m_header.h, m_header.w);
-  short sample;
-  for (int i = 0; i < m_header.h; i++) {
-    for (int j = 0; j < m_header.w; j++) {
-      m_dataFile.read((char *)&sample, sizeof(short));
-      image(i, j, sample);
-    }
-  }
-  m_dataFile.close();
-}
-
 #endif // US_SEQUENCE_READER_3D_H
 #endif // DOXYGEN_SHOULD_SKIP_THIS
